requirements.txt
django
django-admin
djangorestframework
djangorestframework-simplejwt
celery
redis
flower
psycopg2-binary